Refactor: updated structure
