+++ src/components/CurrencySelector.js
