+++ src/index.js
@@ -1,7 +1,8 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
-import "./index.css";
-import App from "./App";
+import "./index.js";
+import App from "./components/App";
+import "./components/App.js";
 
 const root = ReactDOM.createRoot(document.getElementById("root"));
 root.render(
